adding utils file
