+++ client/components/SingleProduct.js
@@ -56,7 +56,11 @@ class SingleProduct extends React.Component {
     const userId = this.props.match.params.userId;
     this.props.deleteProductThunk(this.props.product.id);
     this.setState({...this.state});
-    this.props.history.push(`/${userId}/products`);
+    //console.log(this.props.history)
+    this.props.history.push({
+      pathname: `/${userId}/products`,
+      state: this.props.products
+    });
   }
 
   render() {
@@ -121,6 +125,7 @@ class SingleProduct extends React.Component {
 const mapState = state => ({
   orderId: state.cart.orderId,
   product: state.products.product,
+  products: state.products.products,
   loading: state.products.singleLoading,
   isAdmin: state.user.isAdmin
 });
